requirements.txt

streamlit==1.36.0
streamlit-drawable-canvas==0.9.3
pandas==2.2.2
reportlab==4.2.0
python-docx==1.1.2
pyperclip==1.8.2
Pillow==10.4.0
pytest==8.3.2
